Functional CLI.
