+++ clients/cli/package.json
@@ -13,11 +13,12 @@
   "license": "ISC",
   "dependencies": {
     "@metaplex-foundation/digital-asset-standard-api": "^1.0.0",
-    "@metaplex-foundation/mpl-inscription": "^0.5.1",
+    "@metaplex-foundation/mpl-inscription": "^0.5.2",
     "@metaplex-foundation/mpl-token-metadata": "^3.1.2",
     "@metaplex-foundation/umi": "^0.8.10",
     "@metaplex-foundation/umi-bundle-defaults": "^0.8.10",
     "@solana/web3.js": "^1.87.6",
+    "cli-progress": "^3.12.0",
     "commander": "^11.1.0",
     "p-map": "npm:@esm2cjs/p-map@^5.5.0"
   },
